BrunnerCTF/boot2root_cake-architect-user/requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
playwright==1.54.0
